Fixed project submission. Connected labels to task creation both on b…

+++ backend/src/services/labels/index.js
@@ -1,25 +1,27 @@
 import express from "express";
 import LabelModel from "./scheme.js";
 import createHttpError from "http-errors";
-
+import { requireAuth } from "../../middleware/authMiddleware.js";
 const router = express.Router();
 
 //Create label
-router.post("/", async (req, res, next) => {
+router.post("/", requireAuth, async (req, res, next) => {
   try {
-    const newLabelInput = req.body;
-    const newLabel = new LabelModel(newLabelInput);
-    const { _id } = await newLabel.save();
-    res.send(newLabel);
+      const user = JSON.parse(req.cookies.user)
+      const newLabelInput = req.body;
+      const newLabel = new LabelModel({...newLabelInput, owner: user._id});
+      const savedLabel = await newLabel.save();
+      res.send(savedLabel);
   } catch (error) {
     console.log(error);
     res.status(400).send({ error });
   }
 });
 
-router.get("/", async (req, res, next) => {
+router.get("/", requireAuth, async (req, res, next) => {
   try {
-    const labels = await LabelModel.find();
+    const user = JSON.parse(req.cookies.user)
+    const labels = await LabelModel.find({owner: user._id});
     res.send(labels);
   } catch (error) {
     console.log(error);
@@ -27,29 +29,45 @@ router.get("/", async (req, res, next) => {
   }
 });
 
-router.get("/:id", async (req, res, next) => {
+router.get("/:id", requireAuth, async (req, res, next) => {
   try {
+    const user = JSON.parse(req.cookies.user)
     const label = await LabelModel.findById(req.params.id);
-    res.send(label);
+
+    if(user._id.toString() && label.owner._id.toString()){
+      res.send(label);
+    } else {
+      next(
+        createHttpError(404, "You do not have access to this resource")
+      );
+    }
+
   } catch (error) {
     console.log(error);
     res.status(404).send({error})
   }
 });
 
-router.put("/:id", async (req, res, next) => {
+router.put("/:id", requireAuth, async (req, res, next) => {
   try {
-    const modifiedLabel = await LabelModel.findByIdAndUpdate(
-      req.params.id,
-      req.body,
-      { new: true }
-    );
-    if (modifiedLabel) {
-      res.send(modifiedLabel);
-    }else
-    {
+    const user = JSON.parse(req.cookies.user)
+    const label = await LabelModel.findById(req.params.id);
+    if(user._id.toString() && label.owner._id.toString()){
+      const modifiedLabel = await LabelModel.findByIdAndUpdate(
+        req.params.id,
+        req.body,
+        { new: true }
+      );
+      if (modifiedLabel) {
+        res.send(modifiedLabel);
+      }else{
+        next(
+          createHttpError(404, "label with id: " + req.params.id + " not found")
+        );
+      }
+    } else {
       next(
-        createHttpError(404, "label with id: " + req.params.id + " not found")
+        createHttpError(404, "You do not have permission for this operation")
       );
     }
   } catch (error) {
@@ -57,12 +75,25 @@ router.put("/:id", async (req, res, next) => {
     createHttpError(400, error);
   }
 });
-router.delete("/:id", async (req, res, next) => {
+router.delete("/:id", requireAuth, async (req, res, next) => {
   try {
-    const label = await LabelModel.findByIdAndDelete(req.params.id);
-    res.status(201).send({});
+    const user = JSON.parse(req.cookies.user)
+    const label = await LabelModel.findById(req.params.id);
+    if(user._id.toString() && label.owner._id.toString()){
+      const label = await LabelModel.findByIdAndDelete(req.params.id);
+      res.status(201).send({label});
+    } else {
+      next(
+        createHttpError(404, "You do not have permission for this operation")
+      );
+    }
   } catch (error) {
     console.log(error);
+    next(
+      createHttpError(404, error)
+    );
   }
 });
+
+
 export default router;
